+++ nix/modules/project/settings/default.nix
@@ -75,17 +75,22 @@ in
               });
             }).config;
             cfg = traceSettings name cfg';
+
+            # Some settings must be applied in deterministic order
+            #
+            # NOTE: The `custom` setting must apply first, since it can discard its argument.
             # HACK: buildFromSdist must apply *last*
             # cf. https://github.com/shajra/haskell-flake/pull/252
             # In future, we can refactor this as part of https://github.com/shajra/haskell-flake/issues/285
             # NOTE: removeReferencesTo must apply *before* buildFromSdist, because the
             # later appears it fuck up the former otherwise.
-            impl = builtins.removeAttrs cfg.impl [ "buildFromSdist" "removeReferencesTo" ];
-            fns = lib.attrValues impl ++ [ cfg.impl.buildFromSdist cfg.impl.removeReferencesTo ];
+            # NOTE: separateIntermediatesOutput must apply after buildFromSdist, otherwise
+            # the `installIntermediatesPhase` is set on the `sdist`'s drv instead of the package's.
+            impl = builtins.removeAttrs cfg.impl [ "custom" "buildFromSdist" "removeReferencesTo" "separateIntermediatesOutput" ];
+            fns = [ cfg.impl.custom ] ++ lib.attrValues impl ++ [ cfg.impl.buildFromSdist cfg.impl.removeReferencesTo cfg.impl.separateIntermediatesOutput ];
           in
           lib.pipe super.${name} (
             # TODO: Do we care about the *order* of overrides?
-            # Might be relevant for the 'custom' option.
             lib.concatMap
               (impl: impl)
               fns