+++ CHANGELOG.md
@@ -4,9 +4,14 @@
 
 - Enhancements
   - #382: Support for cabal2nix generated expressions (avoiding IFD for local packages)
+  - Add `extraCabal2nixOptions` and `cabalFlags` to `packages.*`. The former passes custom options to `cabal2nix`. The later does the same, but formatting the flag options appropriately.
   - Support `meta.description` in flake apps. Requires newer version of flake-parts.
   - `settings` module:
     - #384: Add `stan`
+    - #419, #421: Add `installIntermediates`, `previousIntermediates` and `separateIntermediatesOutput`
+    - #427: Add `generateOptparseApplicativeCompletions`
+    - #455: Add `drvAttrs` option for setting arbitrary derivation attributes
+    - #459: Apply `custom` ahead of other settings.
 
 ## 0.5.0 (Jun 24, 2024)